Proyectos Treinta/score_leads_send/pyproject.toml
[tool.poetry]
name = "logysto data"
version = "0.1.0"
description = ""
authors = ["Stivy2 <steven.porras@treinta.co>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
google-cloud-bigquery = "^3.3.2"
google-cloud-secret-manager = "^2.12.4"
pandas = "^1.5.0"
db-dtypes = "^1.0.4"
joblib = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
